Add plot generation from equations feature
This commit includes the addition of a new feature that generates plots from mathematical equations specified by the user. It also contains modifications to the project's configurations files and provides the capability to persist generated plots. Moreover, this update introduces the usage of the FastAPI framework for building the application's backend.